+++ native/android/app/src/main/java/com/notes/MainApplication.java
@@ -33,7 +33,6 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       List<ReactPackage> packages = new PackageList(this).getPackages();
       packages.add(new FxaClientPackage());
-
       return packages;
     }